Fix array access on null literal
This is a follow-up for 79f3c18d02620173da39608b25e9c02b24f92829.

The UseDefChainSimplifier can optimize the IR in a way that it may
contain array access on null literal (like 'null[0] = 1').

However, the backend loses the array type information and then
generates invalid code (using an aput-object for a primitive array
like int[]).

This CL fixes that issue by wrapping the JNullLiteral in a
JReinterpretCast node to keep the array type information.

Note: this cast is only applied to JArrayRef. Indeed JFieldRef and
JMethodCall are managed by the backend by getting the type from the
target field or method.

Bug: 34066768
Change-Id: Id62bf84e336f087bd546cecae843403b7ba6b1aa