Reduce lifetime of CFG
The CfgBuilder schedulable computes the control flow graph (CFG) of a
method (JMethod) and attaches the CFG to the method with a
ControlFlowGraph marker.

With the current plan of schedulables, the CfgBuilder runs on every
method (with code) of every class (to emit in the output) before
being used by the backend (or optional optimizations). Therefore all
these CFGs live in memory at the same time. This is because the
CfgBuilder and the schedulables that need it are in different
subplans.

This CL modifies the location of the CfgBuilder in the plan so that
it is scheduled in the same JMethodAdapter that contains the
schedulables that will process (and potentially modify) the CFG. It
contributes to reducing its lifetime (the JMethodAdapter) so that the
garbage collector can collect it as soon as possible and help to
reduce memory pressure.

Bug: 34494527
Test: 'ant test-jack' with api-v04
Change-Id: I09233c65add456c304b894cf5d9ee425fde943f5