Add a property to force plan amender
The plan amender is only used when there is at least one plugin.
This CL adds the property "jack.force-plan-amender" to force the
plan amender even if there is no plugin.

This is off by default to keep the current behavior of amending the
plan only when loading plugin(s).

Change-Id: I0606af09d59c3d8c0a75cc5bfd981f2964d63548