Annotation does not have generic signature
- Avoid to call writeGenericSignatureMarker for annotation
because an annotation does not have generic type. It could
avoid to Jill to put some information into the Jack IR in
case of mal-formed class file.
- Rename writeOriginalTypeInfoMarker into writeGenericSignatureMarker
because this method write a generic signature marker.
Moreover, no need to emit null when the generic signature
is not valid since markers can be a empty list.

Change-Id: I7f567f68b87e7fd54bb41424ab42b480cde3fd0d